feat: update code check condition to show gauge reward
