+++ src/pages/pool/_id.vue
@@ -36,12 +36,14 @@ import { providePoolStaking } from '@/providers/local/pool-staking.provider';
 import useWeb3 from '@/services/web3/useWeb3';
 import BrandedRedirectCard from '@/components/pool/branded-redirect/BrandedRedirectCard.vue';
 import ExtraRewardCard from '@/components/contextual/pages/pool/ExtraRewardCard.vue';
+import gaugeApi from '@/composables/gaugeReward/gauge.api';
+import useConfig from '@/composables/useConfig';
 /**
  * STATE
  */
 const route = useRoute();
 const poolId = (route.params.id as string).toLowerCase();
-
+const gaugeAddress = ref('');
 /**
  * PROVIDERS
  */
@@ -56,6 +58,7 @@ const { prices } = useTokens();
 const { isWalletReady } = useWeb3();
 const { addAlert, removeAlert } = useAlerts();
 const _isVeBalPool = isVeBalPool(poolId);
+const { networkConfig } = useConfig();
 
 //#region pool query
 const poolQuery = usePoolQuery(poolId, undefined, undefined);
@@ -136,8 +139,25 @@ function addIntersectionObserver(): void {
   observer = new IntersectionObserver(callback, options);
   observer.observe(intersectionSentinel.value);
 }
+async function getGaugeAddress() {
+  try {
+    const response = await gaugeApi.getGaugeAddress({
+      pool_id: poolId,
+      chain_id: networkConfig.chainId,
+    });
+    console.log('🚀 ~ getGaugeAddress ~ response:', response);
+    gaugeAddress.value = response.gauge_address;
+    console.log(
+      '🚀 ~ getGaugeAddress ~  gaugeAddress.value:',
+      gaugeAddress.value
+    );
+  } catch (error) {
+    console.error(error);
+  }
+}
 onMounted(() => {
   addIntersectionObserver();
+  getGaugeAddress();
 });
 onBeforeUnmount(() => {
   observer?.disconnect();
@@ -292,7 +312,7 @@ watch(poolQuery.error, () => {
           />
           <BalLoadingBlock v-if="loadingPool" class="h-40 pool-actions-card" />
           <StakingIncentivesCard
-            v-if="isStakablePool && !loadingPool && pool && isWalletReady"
+            v-if="gaugeAddress && !loadingPool && pool && isWalletReady"
             :pool="pool"
             class="staking-incentives"
           />
@@ -306,7 +326,12 @@ watch(poolQuery.error, () => {
             class="mb-4 h-60 pool-actions-card"
           />
 
-          <ExtraRewardCard v-else :pool="pool" class="mb-4" />
+          <ExtraRewardCard
+            v-else
+            :gaugeAddress="gaugeAddress"
+            :pool="pool"
+            class="mb-4"
+          />
         </BalStack>
       </div>
     </div>
